+++ libs/leads/contracts/jest.config.ts
@@ -1,0 +1,10 @@
+export default {
+  displayName: 'leads-contracts',
+  preset: '../../../jest.preset.js',
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
+  },
+  moduleFileExtensions: ['ts', 'js', 'html'],
+  coverageDirectory: '../../../coverage/libs/leads/contracts',
+};
