+++ libs/leads/contracts/project.json
@@ -1,0 +1,21 @@
+{
+  "name": "leads-contracts",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/leads/contracts/src",
+  "projectType": "library",
+  "tags": [
+    "type:contracts",
+    "scope:leads",
+    "platform:shared",
+    "visibility:internal"
+  ],
+  "targets": {
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "libs/leads/contracts/jest.config.ts"
+      }
+    }
+  }
+}
