+++ libs/leads/data-access/jest.config.ts
@@ -1,0 +1,18 @@
+export default {
+  displayName: 'leads-data-access',
+  preset: '../../../jest.preset.js',
+  testEnvironment: 'jsdom',
+  setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
+  transform: {
+    '^.+\\.(ts|mjs|js|html)$': [
+      'jest-preset-angular',
+      {
+        tsconfig: '<rootDir>/tsconfig.spec.json',
+        stringifyContentPathRegex: '\\.(html|svg)$',
+      },
+    ],
+  },
+  transformIgnorePatterns: ['node_modules/(?!.*\\.mjs$)'],
+  moduleFileExtensions: ['ts', 'html', 'js', 'json', 'mjs'],
+  coverageDirectory: '../../../coverage/libs/leads/data-access',
+};
