+++ libs/leads/data-access/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/leads-api.service';
