+++ libs/leads/feature-flags/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/lead-flags.service';
