+++ libs/leads/feature-onboarding/project.json
@@ -1,0 +1,21 @@
+{
+  "name": "leads-feature-onboarding",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/leads/feature-onboarding/src",
+  "projectType": "library",
+  "tags": [
+    "type:feature",
+    "scope:leads",
+    "platform:web",
+    "visibility:internal"
+  ],
+  "targets": {
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "libs/leads/feature-onboarding/jest.config.ts"
+      }
+    }
+  }
+}
