+++ libs/leads/feature-onboarding/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/lead-onboarding.service';
