+++ libs/leads/feature-topics/project.json
@@ -1,0 +1,21 @@
+{
+  "name": "leads-feature-topics",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/leads/feature-topics/src",
+  "projectType": "library",
+  "tags": [
+    "type:feature",
+    "scope:leads",
+    "platform:web",
+    "visibility:internal"
+  ],
+  "targets": {
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "libs/leads/feature-topics/jest.config.ts"
+      }
+    }
+  }
+}
