+++ libs/leads/feature-topics/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/lead-topics.service';
