+++ tsconfig.base.json
@@ -48,6 +48,21 @@
       "@optimistic-tanuki/finance-ui": ["libs/finance-ui/src/index.ts"],
       "@optimistic-tanuki/hai-ui": ["libs/hai-ui/src/index.ts"],
       "@optimistic-tanuki/leads-ui": ["libs/leads-ui/src/index.ts"],
+      "@optimistic-tanuki/leads-contracts": [
+        "libs/leads/contracts/src/index.ts"
+      ],
+      "@optimistic-tanuki/leads-data-access": [
+        "libs/leads/data-access/src/index.ts"
+      ],
+      "@optimistic-tanuki/leads-feature-flags": [
+        "libs/leads/feature-flags/src/index.ts"
+      ],
+      "@optimistic-tanuki/leads-feature-onboarding": [
+        "libs/leads/feature-onboarding/src/index.ts"
+      ],
+      "@optimistic-tanuki/leads-feature-topics": [
+        "libs/leads/feature-topics/src/index.ts"
+      ],
       "@optimistic-tanuki/fin-commander-data-access": [
         "libs/fin-commander-data-access/src/index.ts"
       ],
